+++ MyUECppProject/Source/MyUECppProject/Component/BoosterComponent.h
@@ -25,17 +25,17 @@ public:
 	virtual void TickComponent(float DeltaTime, ELevelTick TickType, FActorComponentTickFunction* ThisTickFunction) override;
 
 	UPROPERTY(EditAnywhere, BlueprintReadOnly)
-	float FuelMax = 60.0f;
+	double FuelMax = 60.0;
 
 	UPROPERTY(VisibleAnywhere, BlueprintReadWrite)
-	float FuelRest;
+	double FuelRest;
 
 	UPROPERTY(EditAnywhere, BlueprintReadWrite)
 	bool bEnabled = false;
 
 	UPROPERTY(EditAnywhere, BlueprintReadOnly)
-	float Power = 100.0f;
+	double Power = 100.0;
 
 	UFUNCTION(BlueprintPure)
-	float GetVelocity() const;
+	double GetVelocity() const;
 };
